# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "Users_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request — validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — user with this email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create a new user (Admin Only)",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "Users_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive search by user email.",
            "schema": {
              "example": "patient@test.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated list of all users",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResultDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserEntity"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List all users (Admin Only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/patients": {
      "get": {
        "operationId": "Users_findAllPatients",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive search by user email.",
            "schema": {
              "example": "patient@test.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated list of all patients",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResultDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserEntity"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Admin or Doctor role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List all patients (Admin/Doctor Only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/doctors": {
      "get": {
        "operationId": "Users_findAllDoctors",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive search by user email.",
            "schema": {
              "example": "patient@test.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated list of all doctors",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResultDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserEntity"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List all doctors (Admin Only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me/theme": {
      "patch": {
        "operationId": "Users_updateMyTheme",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThemeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Theme preference updated; returns the refreshed user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request — invalid themePreference value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update the current user UI theme preference",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "Users_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID v4)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Admin or Doctor role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found — user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get user detail by ID",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Prescription Management API",
    "description": "API documentation for the MVP Prescription Management System.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
...
      "Role": {
        "type": "string",
        "enum": [
          "ADMIN",
          "DOCTOR",
          "PATIENT"
        ],
        "description": "User role for RBAC"
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code of the error"
          },
          "message": {
            "type": "string",
            "example": "Invalid JSON payload",
            "description": "Error message describing what went wrong"
          },
          "path": {
            "type": "string",
            "example": "/auth/login",
            "description": "The request path that caused the error"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-13T10:30:00.000Z",
            "description": "ISO 8601 timestamp when the error occurred"
          }
        },
        "required": [
          "statusCode",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "ThemePreference": {
        "type": "string",
        "enum": [
          "SYSTEM",
          "LIGHT",
          "DARK"
        ],
        "description": "UI theme preference"
      },
      "DoctorProfileSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "specialty": {
            "type": "string",
            "example": "Cardiology",
            "nullable": true
          },
          "medicalId": {
            "type": "string",
            "example": "MED-12345",
            "nullable": true
          },
          "signatureText": {
            "type": "string",
            "example": "Dr. Jane Doe",
            "nullable": true
          },
          "signatureImageUrl": {
            "type": "string",
            "example": "https://cdn.clinic.com/signatures/jane-doe.png",
            "format": "uri",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
      "PatientProfileSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-05-21T00:00:00.000Z",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current page number (1-indexed)"
          },
          "limit": {
            "type": "number",
            "example": 10,
            "description": "Number of items per page"
          },
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total number of items across all pages"
          },
          "totalPages": {
            "type": "number",
            "example": 5,
            "description": "Total number of pages"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "PaginatedResultDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier (UUID v4)",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "doctor@clinic.com",
            "format": "email"
          },
          "role": {
            "description": "User role for RBAC",
            "example": "DOCTOR",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the user was created",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the user was last updated",
            "example": "2026-01-15T12:30:00.000Z"
          },
          "themePreference": {
            "description": "UI theme preference",
            "example": "SYSTEM",
            "allOf": [
              {
                "$ref": "#/components/schemas/ThemePreference"
              }
            ]
          },
          "doctor": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DoctorProfileSummary"
              }
            ]
          },
          "patient": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PatientProfileSummary"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "createdAt",
          "updatedAt",
          "themePreference"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@clinic.com",
            "description": "User email address",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "***REDACTED-DEV-PASSWORD***",
            "description": "User password (minimum 6 characters)"
          },
          "role": {
            "description": "User role (ADMIN, DOCTOR, PATIENT)",
            "example": "PATIENT",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "specialty": {
            "type": "string",
            "example": "Cardiology",
            "description": "Medical specialty (only applies when role=DOCTOR)"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-05-21",
            "description": "Date of birth ISO-8601 (only applies when role=PATIENT)",
            "format": "date"
          },
          "medicalId": {
            "type": "string",
            "example": "MED-12345",
            "description": "Medical license number (only applies when role=DOCTOR)",
            "maxLength": 64
          },
          "signatureText": {
            "type": "string",
            "example": "Dr. Jane Doe",
            "description": "Text-based handwritten-style signature label rendered on PDFs (DOCTOR only)",
            "maxLength": 120
          },
          "signatureImageUrl": {
            "type": "string",
            "example": "https://cdn.clinic.com/signatures/jane-doe.png",
            "description": "URL to a signature image rendered on PDFs (DOCTOR only)",
            "format": "uri"
          }
        },
        "required": [
          "email",
          "password",
          "role"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier (UUID v4)",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "doctor@clinic.com",
            "format": "email"
          },
          "role": {
            "description": "User role for RBAC",
            "example": "DOCTOR",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the user was created",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the user was last updated",
            "example": "2026-01-15T12:30:00.000Z"
          },
          "themePreference": {
            "description": "UI theme preference",
            "example": "SYSTEM",
            "allOf": [
              {
                "$ref": "#/components/schemas/ThemePreference"
              }
            ]
          },
          "doctor": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DoctorProfileSummary"
              }
            ]
          },
          "patient": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PatientProfileSummary"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "createdAt",
          "updatedAt",
          "themePreference"
        ]
      },
      "UpdateThemeDto": {
        "type": "object",
        "properties": {
          "themePreference": {
            "example": "DARK",
            "description": "UI theme preference for the current user. SYSTEM follows the OS preference.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ThemePreference"
              }
            ]
          }
        },
        "required": [
          "themePreference"
        ]
      },
...
    }
  }
}